+++ MODULE_GLOB.mjs
@@ -1,8 +1,8 @@
 // @ts-check
 
 /**
- * File glob pattern to recursively match TypeScript (`.mts`, and `.cts`;
- * excluding TypeScript definition files) and JavaScript (`.mjs`, `.cjs`, and
- * `.js`) modules.
+ * File glob pattern to recursively match TypeScript (`.mts`, `.cts`, `.ts`, and
+ * `.tsx`; excluding TypeScript definition files) and JavaScript (`.mjs`,
+ * `.cjs`, and `.js`) modules.
  */
-export default "**/{!(*.d).mts,!(*.d).cts,*.{mjs,cjs,js}}";
+export default "**/{!(*.d).mts,!(*.d).cts,!(*.d).ts,*.{mjs,cjs,js,tsx}}";